+++ engine/modernizers/cobol_system/rules/data_definition_rules.py
@@ -7,7 +7,7 @@ to Python variable declarations.
 
 from typing import Dict, Any, List, Optional
 from .base_rule import BaseRule
-from ..parser.cobol_lst import LosslessNode
+from ..parsers.cobol_lst import LosslessNode
 
 
 class DataDescriptionRule(BaseRule):
