+++ engine/modernizers/cobol_system/rules/rule_engine.py
@@ -32,7 +32,7 @@ from .arithmetic_rules import (
     ComputeRule,
     ComparisonRule
 )
-from ..parser.cobol_lst import LosslessNode
+from ..parsers.cobol_lst import LosslessNode
 
 
 class RuleEngine:
